pyproject.toml
[project]
name = "agent2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.3.0",
    "faiss-cpu>=1.12.0",
    "flagembedding>=1.3.5",
    "huggingface-hub>=0.36.0",
    "langchain>=1.0.2",
    "langchain-openai>=1.0.1",
    "numpy>=2.2.6",
    "openai>=2.6.1",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.7",
    "ragas>=0.3.1",
    "sentence-transformers>=5.1.2",
    "tiktoken>=0.12.0",
]
